requirements.txt
aiohttp
aiosocksy
rsa
requests[socks]
configobj
pyinstaller
flask
tailer
python-telegram-bot